[fix] RDF: Add taxon search UI, in API [fix] RDF: Add equivalence to RDF model [fix] RDF: Fix sparql endpoint, and allow to execute on a remote service [fix] RDF: Download Sandre+MNHN taxons into local triple store

+++ sumaris-core-rdf/src/main/java/net/sumaris/rdf/dao/RdfModelDaoImpl.java
@@ -105,7 +105,7 @@ public class RdfModelDaoImpl extends HibernateDaoSupport implements RdfModelDao 
 
         String hql = getSelectHqlQuery(domain, className, page.getSortAttribute(), page.getSortDirection());
         TypedQuery<T> typedQuery = entityManager.createQuery(hql, aClass)
-                .setFirstResult(page.getOffset())
+                .setFirstResult((int)page.getOffset())
                 .setMaxResults(page.getSize());
 
         // When using fetch join, stream are not supported, so use a list
